스파르타 과제/3. Django, DRF/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
joblib==1.4.2
mypy-extensions==1.0.0
mysqlclient==2.2.7
numpy==2.1.3
openai==1.57.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
